# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Experimental quantum-resistant, hardware-aware auth service.",
        "title": "Quantum Auth API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1042",
    "basePath": "/",
    "paths": {
        "/api/secure-ping": {
            "get": {
                "description": "Confirms that the request was authenticated using TPM + PQ signatures.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secure"
                ],
                "summary": "Quantum-secured ping",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.SecurePingResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/challenge": {
            "post": {
                "description": "Issues a short-lived challenge (nonce) for a registered device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Issue auth challenge",
                "parameters": [
                    {
                        "description": "Challenge request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.authChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.authChallengeResponse"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "device not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "description": "Verifies a signed challenge response using TPM and PQ signatures plus password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify auth response (hybrid TPM + PQ)",
                "parameters": [
                    {
                        "description": "Verify request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.authVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.authVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "invalid signature or password",
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.authVerifyResponse"
                        }
                    },
                    "404": {
                        "description": "challenge or device not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/devices/register": {
            "post": {
                "description": "Registers a new authentication device (TPM + PQ keys) for an existing user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Register device",
                "parameters": [
                    {
                        "description": "Device registration",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.registerDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.registerDeviceResponse"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Create a user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Signup",
                "parameters": [
                    {
                        "description": "Signup payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quantumhttp.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "quantumhttp.SecurePingResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.SignupRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.authChallengeRequest": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.authChallengeResponse": {
            "type": "object",
            "properties": {
                "challenge_id": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.authVerifyRequest": {
            "type": "object",
            "properties": {
                "challenge_id": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "pq_signature": {
                    "type": "string"
                },
                "tpm_signature": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.authVerifyResponse": {
            "type": "object",
            "properties": {
                "authenticated": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.registerDeviceRequest": {
            "type": "object",
            "properties": {
                "device_label": {
                    "type": "string"
                },
                "pq_public_key": {
                    "type": "string"
                },
                "tpm_public_key": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                }
            }
        },
        "quantumhttp.registerDeviceResponse": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                }
            }
        }
    }
}